requirements.txt
selenium==4.27.1
python-dotenv==1.0.1
bs4
requests
python-docx
loguru